Refactor newloader Qwen3 runtime onto GptModelBase (#1195)
* refactor(loader): unify newloader runtime on GptModelBase

* fix(loader): preserve shared GptModelBase contracts
