+++ rtp_llm/models_py/BUILD
@@ -117,6 +117,7 @@ py_library(
     srcs = [
         "__init__.py",
         "model_loader.py",
+        "model_desc/module_base.py",
         "module_base.py",
         "registry.py",
         "weight_mapper.py",
