+++ rtp_llm/models_py/modules/factory/attention/attn_factory.py
@@ -1,7 +1,8 @@
-import logging
-from typing import Callable, Dict, List, Optional, Union
+from __future__ import annotations
 
-from rtp_llm.model_loader.model_weight_info import ModelWeights
+import logging
+from typing import TYPE_CHECKING, Callable, Dict, List, Optional, Union
+
 from rtp_llm.models_py.modules.factory.attention.fmha_impl_base import (
     FMHAImplBase,
     MlaImplBase,
@@ -18,6 +19,9 @@ from rtp_llm.utils.model_weight import W
 
 AttentionImpl = Union[FMHAImplBase, MlaImplBase]
 AttentionImplFactory = Callable[..., AttentionImpl]
+
+if TYPE_CHECKING:
+    from rtp_llm.model_loader.model_weight_info import ModelWeights
 
 # Lists to store registered implementations
 PREFILL_MHA_IMPS: List[type[FMHAImplBase]] = []
